Request添加Post支持

+++ Crocodile/App/Client/User/UserDetailRequest.h
@@ -9,6 +9,6 @@
 
 @interface UserDetailRequest: NSObject<Request>
 
-- (instancetype)initWithUID:(NSInteger)uid;
+- (instancetype)initWithUID:(int)uid;
 
 @end
